Add SuggestedProducts and fix layout
