+++ messages/context.json
@@ -7,5 +7,6 @@
   "store/shelf.buy-together.change.label": "Change label",
   "store/shelf.buy-together.remove.label": "Remove label",
   "store/shelf.buy-together.total-products.label": "Total products label",
-  "store/shelf.refresh.refresh-product-summary.title": "Refresh Product Summary title"
+  "store/shelf.refresh.refresh-product-summary.title": "Refresh Product Summary title",
+  "store/shelf.refresh.suggested-products.title": "Refresh Suggested Products title"
 }
