+++ messages/en.json
@@ -7,5 +7,6 @@
   "store/shelf.buy-together.change.label": "Change",
   "store/shelf.buy-together.remove.label": "Remove",
   "store/shelf.buy-together.total-products.label": "Buy {total} {total, plural, one {product} other {products}} for",
-  "store/shelf.refresh.refresh-product-summary.title": "You saw"
+  "store/shelf.refresh.refresh-product-summary.title": "You saw",
+  "store/shelf.refresh.suggested-products.title": "Suggested for you"
 }
