+++ messages/es.json
@@ -7,5 +7,6 @@
   "store/shelf.buy-together.change.label": "Cambiar",
   "store/shelf.buy-together.remove.label": "Eliminar",
   "store/shelf.buy-together.total-products.label": "Comprar {total} {total, plural, one {producto} other {productos}} por",
-  "store/shelf.refresh.refresh-product-summary.title": "Has visto"
+  "store/shelf.refresh.refresh-product-summary.title": "Has visto",
+  "store/shelf.refresh.suggested-products.title": "Sugeridos para ti"
 }
