+++ react/typings/vtex.slider-layout.d.ts
@@ -1,0 +1,3 @@
+declare module 'vtex.slider-layout' {
+  export const SliderLayout
+}
